Inclusao de regras exclusao de agendamento

+++ api/Domains/UseCase/AppointmentService/AppointmentDeleteService.cs
@@ -1,5 +1,6 @@
 using System.Threading.Tasks;
 using Domains.Repositories;
+using Domains.UseCase.AppointmentService.Exceptions;
 using Entities;
 using Shared.Exceptions;
 
@@ -20,6 +21,7 @@ namespace Domains.UseCase.AppointmentService
             if(id == 0) throw new NotFoundRegisterException("Agendamento não Encontrado.");
             var appointment = await _repository.FindById(id);
             if(appointment == null ) throw new NotFoundRegisterException("Agendamento não Encontrado.");
+            if(appointment.DateTimeCollected != null && appointment.DateTimeDelivery == null ) throw new AppointmentNotExcludeException("Carro possui agendamento em aberto. Verifique para exclusão.");
             await _repository.Delete(appointment);
         }
     }
